Implement OperationQueue for managing asynchronous operations in State class; refactor get() and set() methods to utilize the new queue system
