+++ packages/protocol/src/state/State.ts
@@ -36,108 +36,10 @@ export class WithStateServiceProvider {
   }
 }
 
-type AsyncOperation<T> = () => Promise<T>;
-
 /**
  * Utilities for runtime module state, such as get/set
  */
 export class State<Value> extends Mixin(WithPath, WithStateServiceProvider) {
-  private queue: Array<{
-    operation: AsyncOperation<any>;
-    resolve: (value: any) => void;
-    reject: (reason: any) => void;
-  }> = [];
-
-  private running = false;
-
-  private pendingCount = 0;
-
-  private drainResolvers: Array<() => void> = [];
-
-  // Enqueue an operation and return a promise for its result
-  queueOperation<T>(operation: AsyncOperation<T>): Promise<T> {
-    console.log("queuing operation, previous pendingCount:", this.pendingCount);
-    return new Promise<T>((resolve, reject) => {
-      this.queue.push({
-        operation,
-        resolve,
-        reject,
-      });
-      this.pendingCount += 1;
-      console.log("pendingCount:", this.pendingCount);
-      // If not running, start processing the queue
-      if (!this.running) {
-        this.running = true;
-        // eslint-disable-next-line @typescript-eslint/no-floating-promises
-        this.runNext();
-      }
-    });
-  }
-
-  // Run the next operation in the queue
-  private async runNext() {
-    // If there are no operations left, mark as not running and possibly resolve drain
-    if (this.queue.length === 0) {
-      console.log("no more operations to resolve");
-      this.running = false;
-      // If everything is done, resolve any drain promises
-      if (this.pendingCount === 0) {
-        this.resolveDrain();
-      }
-      return;
-    }
-
-    console.log("running operation");
-
-    const { operation, resolve, reject } = this.queue.shift()!;
-    try {
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-      const result = await operation();
-      console.log("operation resolved");
-      resolve(result);
-    } catch (error) {
-      reject(error);
-    } finally {
-      this.pendingCount -= 1;
-      console.log("pendingCount:", this.pendingCount);
-      console.log("queue length:", this.queue.length);
-      if (this.queue.length > 0) {
-        console.log("running next operation");
-        // eslint-disable-next-line @typescript-eslint/no-floating-promises
-        this.runNext();
-      } else {
-        // No more tasks in the queue
-        console.log("no more operations to resolve");
-        this.running = false;
-        if (this.pendingCount === 0) {
-          console.log("no more pending operations");
-          this.resolveDrain();
-        }
-      }
-    }
-  }
-
-  // Returns a promise that resolves when all queued operations have completed
-  onCompleted() {
-    if (this.queue.length === 0 && this.pendingCount === 0) {
-      // No pending tasks, resolve immediately
-      return Promise.resolve();
-    }
-    console.log("waiting for drain, pendingCount:", this.pendingCount);
-    return new Promise<void>((resolve) => {
-      this.drainResolvers.push(resolve);
-    });
-  }
-
-  private resolveDrain() {
-    if (this.queue.length === 0 && this.pendingCount === 0) {
-      while (this.drainResolvers.length > 0) {
-        const resolver = this.drainResolvers.shift()!;
-        resolver();
-      }
-    }
-  }
-
   /**
    * Creates a new state wrapper for the provided value type.
    *
@@ -229,19 +131,21 @@ export class State<Value> extends Mixin(WithPath, WithStateServiceProvider) {
    * @returns Option representation of the current state.
    */
   public async get(): Promise<Option<Value>> {
-    return await this.queueOperation(async () => {
-      const option = await this.witnessFromState();
+    return await container
+      .resolve(RuntimeMethodExecutionContext)
+      .operationQueue.queueOperation(async () => {
+        const option = await this.witnessFromState();
 
-      this.hasPathOrFail();
+        this.hasPathOrFail();
 
-      const stateTransition = StateTransition.from(this.path, option);
+        const stateTransition = StateTransition.from(this.path, option);
 
-      container
-        .resolve(RuntimeMethodExecutionContext)
-        .addStateTransition(stateTransition);
+        container
+          .resolve(RuntimeMethodExecutionContext)
+          .addStateTransition(stateTransition);
 
-      return option;
-    });
+        return option;
+      });
   }
 
   /**
@@ -256,22 +160,24 @@ export class State<Value> extends Mixin(WithPath, WithStateServiceProvider) {
    * @param value - Value to be set as the current state
    */
   public async set(value: Value) {
-    await this.queueOperation(async () => {
-      // link the transition to the current state
-      const fromOption = await this.witnessFromState();
-      const toOption = Option.fromValue(value, this.valueType);
+    await container
+      .resolve(RuntimeMethodExecutionContext)
+      .operationQueue.queueOperation(async () => {
+        // link the transition to the current state
+        const fromOption = await this.witnessFromState();
+        const toOption = Option.fromValue(value, this.valueType);
 
-      this.hasPathOrFail();
+        this.hasPathOrFail();
 
-      const stateTransition = StateTransition.fromTo(
-        this.path,
-        fromOption,
-        toOption
-      );
+        const stateTransition = StateTransition.fromTo(
+          this.path,
+          fromOption,
+          toOption
+        );
 
-      container
-        .resolve(RuntimeMethodExecutionContext)
-        .addStateTransition(stateTransition);
-    });
+        container
+          .resolve(RuntimeMethodExecutionContext)
+          .addStateTransition(stateTransition);
+      });
   }
 }
